Make distinction between version numbers.   - gregoriotex version of style xxxxxxxx     Changing this style requires changing the gregoriotex plugins.   - gregorio version of style x.x.x
