+++ VERSION
@@ -1,1 +1,2 @@
-2.4.2
+gregoriotex_api_version = 20150216
+gregorio_version = 2.4.1
