+++ tex/gregoriotex-signs.tex
@@ -18,7 +18,7 @@
 % this file contains definitions of signs (bar, episemus, punctum, alterations)
 
 \gredeclarefileversion{gregoriotex-signs.tex}%
- {\directlua{tex.write(gregoriotex.get_greversion())}}
+ {\directlua{tex.write(gregoriotex.get_greapiversion())}}
 
 \def\greusestylecommon{%
   \ifnum\greusestylefont=0\relax %
