+++ tex/gregoriotex-spaces.tex
@@ -17,7 +17,7 @@
 % this file contains definitions of spaces
 
 \gredeclarefileversion{gregoriotex-spaces.tex}%
- {\directlua{tex.write(gregoriotex.get_greversion())}}
+ {\directlua{tex.write(gregoriotex.get_greapiversion())}}
 
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 %% macros for tuning penalties
