+++ tex/gregoriotex-syllable.tex
@@ -17,7 +17,7 @@
 % this file contains definitions of the glyphs and the syllables
 
 \gredeclarefileversion{gregoriotex-syllable.tex}%
- {\directlua{tex.write(gregoriotex.get_greversion())}}
+ {\directlua{tex.write(gregoriotex.get_greapiversion())}}
 
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 %% macros for the typesetting of the different glyphs
