+++ tex/gregoriotex-symbols.tex
@@ -17,7 +17,7 @@
 % this file contains definitions of the symbols
 
 \gredeclarefileversion{gregoriotex-symbols.tex}%
- {\directlua{tex.write(gregoriotex.get_greversion())}}
+ {\directlua{tex.write(gregoriotex.get_greapiversion())}}
 
 % a temp counter
 \newcount\gretempfactor %
